Fix Field Formatter mis-formatting
Previously, if a field were at the start of a paragraph and the
immediately preceding paragraph were, e.g., a section header
with non-basic formatting, then the non-basic formatting would
be applied to the field, undesirably.

This makes the formatting search happen in the *forward* direction
if the field is at the start of a paragraph.